+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Always run `yarn` locally as an initial step of `vtex link` and `vtex publish` for `./node` and `./react` folders
 
 ## [2.56.4] - 2019-05-10
 ### Fixed